Undid a lot of side changes I made to others.
As I was learning this game and prototyping tournament systems for it
I was getting a grips for what made sides good and bad, and tried to
figure that out by either fixing bugs on other sides or correcting flaws
or sometimes just rewriting them. Ultimately it's better that I mark
any work I have done as mine, and flag them appropriately. I have
undone all of my changes except for name disambiguation and formatting
changes.